JPA and other things

+++ BackEnd-portafolio/src/main/java/backendportafolio/repository/contracts/IUserCredentialRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IUserCredentialRepository extends JpaRepository<EstudiantesEntity,Long > {
+public interface IUserCredentialRepository extends JpaRepository<EstudiantesEntity,Integer > {
 }
